moved uthash.h include to pio_internal.h
