+++ src/clib/pio.h
@@ -14,7 +14,6 @@
 #include <stdbool.h>
 #include <string.h> /* memcpy */
 #include <mpi.h>
-#include <uthash.h>
 
 #include <netcdf.h>
 
@@ -96,34 +95,6 @@
 
 /** A convience macro for netCDF integration code. */
 #define NC_PIO NC_UDF0
-
-/**
- * IO region structure.
- *
- * Each IO region is a unit of data which can be described using start
- * and count arrays. Each IO task may in general have multiple io
- * regions per variable.  The box rearranger will have at most one io
- * region per variable.
- *
- * The write from a particular IO task is divided into 1 or more
- * regions each of which can be described using start and count. The
- * io_region typedef is a linked list of those regions.
- */
-typedef struct io_region
-{
-    /** The offset from the beginning of the data buffer to the
-     * beginning of this region.  */
-    int loffset;
-
-    /** Start array for this region. */
-    PIO_Offset *start;
-
-    /** Count array for this region. */
-    PIO_Offset *count;
-
-    /** Pointer to the next io_region in the list. */
-    struct io_region *next;
-} io_region;
 
 /**
  * Rearranger comm type. The rearranger option values must match the
